requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-crispy-forms==2.1
crispy-bootstrap5==0.7
Pillow>=9.0.0
django-extensions==3.2.3
django-imagekit==5.0.0
django-environ==0.11.2
stripe==7.8.0
django-ratelimit==4.1.0
django-debug-toolbar==4.2.0
django-countries==7.5.1
django-phonenumber-field==7.2.0
phonenumbers==8.13.25
django-taggit==4.0.0
django-allauth==0.57.0
django-ckeditor==6.7.0
django-bootstrap5==23.3